requirements.txt
# CoreSense AI Platform - Clean Requirements
# Optimized dependencies following DRY and YAGNI principles

# Core Framework
streamlit>=1.28.0
python-dotenv>=1.0.0

# Data Processing (only if needed)
numpy>=1.24.0

# Async Support
asyncio-compat>=0.1.3

# HTTP Requests
httpx>=0.25.0
aiohttp>=3.8.0  # For coral client async HTTP requests

# Database (SQLite by default, PostgreSQL optional)
sqlalchemy>=2.0.0

# Authentication & Security
bcrypt>=4.0.0
python-jose[cryptography]>=3.3.0
passlib>=1.7.4

# Email (simplified)
yagmail>=0.15.0

# Development & Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Optional: Advanced features (comment out if not needed)
openai>=1.100.0   # AI features used in coral orchestration
mcp>=0.5.0        # MCP servers are part of architecture
plotly>=5.17.0    # Charts needed for muscle activation display
# pandas>=2.0.0   # Only if data analysis is needed

# Production deployment
gunicorn>=21.2.0